// src/shared/interfaces.ts
export interface SDKs {
    readonly dart?: string;
    readonly dartVersion?: string;
    readonly flutter?: string;
    readonly flutterVersion?: string;
    readonly isDartSDKFromFlutter: boolean;
}

export interface DartSDKs extends SDKs {
    readonly dart: string;
}

export interface FlutterSDKs extends SDKs {
    readonly flutter: string;
}

export interface DartWorkspaceContext extends WorkspaceContext {
    readonly sdks: DartSDKs;
}